Extract runtime startup orchestration support
